Plot intensity distribution of metabolite across conditions

Plot intensity distribution of metabolite across conditions

Usage

compare_metabo_distr(obj, metabolite, conds_of_interest = NULL)

Arguments

obj

A S2IsoMEr object after enrichment analysis.

metabolite

A character indicating metabolite name

conds_of_interest

Optional. A character vector specifying which conditions to include in the plot. If NULL (the default), all conditions are included.

Value

A ggplot2 object displaying the density plot of metabolite intensities across conditions.

Examples

compare_metabo_distr(obj, metabolite="C6H8N2O3-H")
#> Error in obj$scmatrix[metabolite, ] %>% as.data.frame(): could not find function "%>%"