Example Multi condition MSEA results

The example_MSEA_multicond dataset is a data frame that contains results of running Run_bootstrap_MSEA on multiple pairwise conditions compared to same reference.

Usage

data(example_MSEA_multicond)

Format

A data frame with 100 rows and 9 variables:

Term

Character: Term name of metabolite set (e.g. Triacylglycerols)

n

Numeric: Median term/query overlap size over bootstraps.

ES_median

Numeric: Median of NES (normalized enrichment score) for fgsea or ES for ks_signed

ES_sd

Numeric: stadndard deviation of NES (normalized enrichment score) for fgsea or ES for ks_signed

p.value_combined

Numeric: Combined p-value using metap::sumlog

q.value_combined

Numeric: Combined adjusted p-value using metap::sumlog

fraction.bootstrap.presence

Numeric: Proportion of bootstraps the given term was tested

condition.x

Character: Reference condition

condition.y

Character: Query condition

Examples

# Load the example data frame
data(example_MSEA_multicond)

# View the first few rows
head(example_MSEA_multicond)
#> # A tibble: 6 × 9
#>   Term                      n ES_median  ES_sd p.value_combined q.value_combined
#>   <chr>                 <dbl>     <dbl>  <dbl>            <dbl>            <dbl>
#> 1 Triacylglycerols         17      1.82 0.0302         1.39e-53         2.00e-39
#> 2 Glycerophosphocholin…    18     -2.41 0.230          4.93e-54         4.38e-39
#> 3 Diradylglycerols          6      1.50 0.0737         3.31e-22         4.83e-19
#> 4 Glycerophosphoethano…     7     -1.36 0.186          5.70e-12         2.13e-11
#> 5 Triacylglycerols         10      2.19 0.0280         5.07e-84         8.31e-60
#> 6 Diradylglycerols          6      1.76 0.128          1.84e-50         3.07e-39
#> # ℹ 3 more variables: fraction.bootstrap.presence <dbl>, condition.x <chr>,
#> #   condition.y <chr>